43 - Cleaning Up Project CRUD

+++ src/projects/views.py
@@ -7,6 +7,7 @@ from django.shortcuts import redirect, render, get_object_or_404
 
 from . import forms
 
+PROJECT_CAN_DELETE_ITEM_THRESHOLD = 2
 
 @login_required
 def project_list_view(request):
@@ -16,6 +17,7 @@ def project_list_view(request):
 @login_required
 def project_detail_update_view(request, handle=None):
     instance = get_object_or_404(Project, handle=handle, owner=request.user)
+    items_qs = instance.item_set.all()
     form = forms.ProjectUpdateForm(request.POST or None, instance=instance)
     if form.is_valid():
         project_obj = form.save(commit=False)
@@ -24,6 +26,7 @@ def project_detail_update_view(request, handle=None):
         return redirect(project_obj.get_absolute_url())
     context = {
         "instance": instance,
+        "items_qs": items_qs,
         "form": form,
     }
     return render(request, "projects/detail.html", context)
@@ -31,7 +34,13 @@ def project_detail_update_view(request, handle=None):
 @login_required
 def project_delete_view(request, handle=None):
     instance = get_object_or_404(Project, handle=handle, owner=request.user)
+    items_qs = instance.item_set.all()
+    items_count = items_qs.count()
+    items_exists = items_qs.exists()
     if request.method == "POST":
+        if items_exists and items_count >= PROJECT_CAN_DELETE_ITEM_THRESHOLD:
+            messages.error(request, f"Cannot delete project with {PROJECT_CAN_DELETE_ITEM_THRESHOLD} or more active items")
+            return redirect(instance.get_delete_url())
         instance.delete()
         return redirect("projects:list")
     return render(request, "projects/delete.html", {"instance": instance})
@@ -45,7 +54,7 @@ def project_create_view(request):
     form = forms.ProjectCreateForm(request.POST or None)
     if form.is_valid():
         project_obj = form.save(commit=False)
-        project_obj.project = request.project
+        project_obj.owner = request.user
         project_obj.added_by = request.user 
         project_obj.save()
         return redirect(project_obj.get_absolute_url())
@@ -73,7 +82,7 @@ def activate_project_view(request, handle=None):
         messages.error(request, "Project could not activate. try again.")
         return redirect("/projects")
     request.session['project_handle'] = handle
-    messages.error(request, "Project activated.")
+    messages.success(request, "Project activated.")
     return redirect("/")
